+++ CHANGELOG.md
@@ -6,7 +6,8 @@ LUYA CHANGELOG
 
 **ISSUES**
 
-- `#687` fixed issue in StorageImporter and empty storage folder.
+- `#691` Create the module migration folder if the folder does not exists.
+- `#687` Fixed issue in StorageImporter and empty storage folder.
 - `#684` Improved styling of error api module mails.
 
 1.0.0-beta3 (4.1.2016)